requirements.txt
huggingface_hub
python-dotenv
pydantic
pdf2image
Pillow
streamlit
